refactor: consolidate agent transform logic (#196)
