+++ packages/shared/src/index.ts
@@ -4,3 +4,4 @@ export * from './sanitize';
 export * from './api-helpers';
 export * from './hashtags';
 export * from './mentions';
+export * from './transforms';
