+++ packages/shared/src/transforms/index.ts
@@ -1,0 +1,2 @@
+export { transformAgent, transformAuthor } from './agent';
+export type { AgentResponse, AuthorResponse } from './agent';
